Add a dry run setting for metadata import.

+++ kuha/importer/harvest.py
@@ -3,7 +3,7 @@ import logging
 from .. import models
 from ..exception import HarvestError
 
-def update(provider, since=None, purge=False):
+def update(provider, since=None, purge=False, dry_run=False):
     """Update metadata formats, items, records and sets.
 
     Parameters
@@ -38,6 +38,9 @@ def update(provider, since=None, purge=False):
     purge: bool
         If `True`, purge deleted formats, items and records from the
         database.
+    dry_run: bool
+        If `True`, fetch records as usual but do not actually change the
+        database.
 
     Raises
     ------
@@ -45,14 +48,14 @@ def update(provider, since=None, purge=False):
         If the provider raises an exception and the harvest cannot be
         continued.
     """
-    update_formats(provider, purge)
-    update_items(provider, purge)
-    update_records(provider, since)
+    prefixes = update_formats(provider, purge, dry_run)
+    identifiers = update_items(provider, purge, dry_run)
+    update_records(provider, identifiers, prefixes, since, dry_run)
 
 
-def update_formats(provider, purge=False):
+def update_formats(provider, purge=False, dry_run=False):
     log = logging.getLogger(__name__)
-    log.info('Updating metadata formats...')
+    log.debug('Updating metadata formats...')
 
     try:
         new_formats = provider.formats()
@@ -68,16 +71,18 @@ def update_formats(provider, purge=False):
         removed = 0
         for prefix, format_ in old_formats.iteritems():
             if prefix not in new_formats:
-                format_.mark_as_deleted()
+                if not dry_run:
+                    format_.mark_as_deleted()
                 removed += 1
 
         added = 0
         for prefix, (namespace, schema) in new_formats.iteritems():
-            models.Format.create_or_update(prefix, namespace, schema)
+            if not dry_run:
+                models.Format.create_or_update(prefix, namespace, schema)
             if prefix not in old_formats:
                 added += 1
 
-        if purge:
+        if purge and not dry_run:
             models.purge_deleted()
     except Exception as e:
         models.rollback()
@@ -85,7 +90,10 @@ def update_formats(provider, purge=False):
         raise HarvestError(e.message)
 
     else:
-        models.commit()
+        if dry_run:
+            models.rollback()
+        else:
+            models.commit()
         # TODO: log number of changed formats
         log.info(
             'Removed {0} format{1} and added {2} format{3}.'
@@ -95,10 +103,12 @@ def update_formats(provider, purge=False):
             )
         )
 
+        return new_formats.keys()
 
-def update_items(provider, purge=False):
+
+def update_items(provider, purge=False, dry_run=False):
     log = logging.getLogger(__name__)
-    log.info('Looking for added and removed items...')
+    log.debug('Looking for added and removed items...')
 
     try:
         new_identifiers = frozenset(map(unicode, provider.identifiers()))
@@ -111,25 +121,30 @@ def update_items(provider, purge=False):
         removed = 0
         for identifier, item in old_items.iteritems():
             if identifier not in new_identifiers:
-                item.mark_as_deleted()
+                if not dry_run:
+                    item.mark_as_deleted()
                 log.debug('deleted {0}'.format(identifier))
                 removed += 1
 
         added = 0
         for identifier in new_identifiers:
-            models.Item.create_or_update(identifier)
+            if not dry_run:
+                models.Item.create_or_update(identifier)
             if identifier not in old_items:
                 log.debug('added {0}'.format(identifier))
                 added += 1
 
-        if purge:
+        if purge and not dry_run:
             models.purge_deleted()
     except Exception as e:
         models.rollback()
         log.exception('Failed to update items: {0}'.format(e))
         raise HarvestError(e.message)
     else:
-        models.commit()
+        if dry_run:
+            models.rollback()
+        else:
+            models.commit()
         log.info(
             'Removed {0} item{1} and added {2} item{3}.'
             ''.format(
@@ -138,14 +153,18 @@ def update_items(provider, purge=False):
             )
         )
 
+        return new_identifiers
 
-def update_sets(provider, identifier):
+
+def update_sets(provider, identifier, dry_run=False):
     log = logging.getLogger(__name__)
     log.debug('Updating sets...')
 
     # Remove the item from old sets.
-    item = models.Item.get(identifier)
-    item.clear_sets()
+    item = None
+    if not dry_run:
+        item = models.Item.get(identifier)
+        item.clear_sets()
 
     sets = provider.get_sets(identifier)
     if len(sets) == 0:
@@ -154,21 +173,22 @@ def update_sets(provider, identifier):
     sets.sort(key=lambda (spec, _): spec.count(u':'))
     # TODO: make sure that sets contain the parent sets of all sets
     for spec, name in sets:
-        set_ = models.Set.create_or_update(spec, name)
-        item.add_to_set(set_)
+        if not dry_run:
+            set_ = models.Set.create_or_update(spec, name)
+            item.add_to_set(set_)
 
 
-def update_records(provider, since=None):
+def update_records(provider,
+                   identifiers,
+                   prefixes,
+                   since=None,
+                   dry_run=False):
     log = logging.getLogger(__name__)
     if since is not None:
         log.info('Updating records modified since {0} UTC...'
                  ''.format(since))
     else:
         log.info('Updating all records...')
-
-    prefixes = [f.prefix for f in models.Format.list(ignore_deleted=True)]
-    identifiers = [i.identifier
-                   for i in models.Item.list(ignore_deleted=True)]
 
     updated = 0
     for identifier in identifiers:
@@ -179,7 +199,7 @@ def update_records(provider, since=None):
                 continue
             log.debug('Updating item "{0}"'.format(identifier))
 
-            update_sets(provider, identifier)
+            update_sets(provider, identifier, dry_run)
         except Exception as e:
             log.exception(
                 'Failed to update item "{0}": {1}'
@@ -190,11 +210,13 @@ def update_records(provider, since=None):
             try:
                 xml = provider.get_record(identifier, prefix)
                 if xml is None:
-                    models.Record.mark_as_deleted(identifier, prefix)
+                    if not dry_run:
+                        models.Record.mark_as_deleted(identifier, prefix)
                 else:
-                    models.Record.create_or_update(
-                        identifier, prefix, xml
-                    )
+                    if not dry_run:
+                        models.Record.create_or_update(
+                            identifier, prefix, xml
+                        )
                     updated += 1
             except Exception as e:
                 models.rollback()
@@ -205,7 +227,10 @@ def update_records(provider, since=None):
             else:
                 # Commit after each record so that the (esp. SQLite)
                 # database does not get locked for a long time.
-                models.commit()
+                if dry_run:
+                    models.rollback()
+                else:
+                    models.commit()
                 log.debug('Processed item "{0}"'.format(identifier))
 
     # End the transaction in case no records were updated.
